Revert "Refactor AppRender Longs attributes"
This reverts commit 9dd4723a8cb9e998212584fd4843655b474a3dc5.

+++ src/commonMain/kotlin/com/yesferal/hornsapp/core/data/repository/SettingsRepositoryImpl.kt
@@ -8,7 +8,7 @@ class SettingsRepositoryImpl(
     private val environmentDataSource: EnvironmentDataSource,
     private val onBoardingDataSource: OnBoardingDataSource
 ) : SettingsRepository {
-    override val screenDelay: Int
+    override val screenDelay: Long
         get() = 300
 
     override fun getEnvironments() = environmentDataSource.getEnvironments()
